Перенес код освещения в отдельный шейдер. Изменил код загрузки данных текстур, теперь текстуры загружаются в видеопамять в 16 бит (RGB5-A1). Оптимизировал функции отрисовки. Уменьшение объема занимаемой клиентом памяти и увеличение производительности. Заготовка для пробы использования VBO.

+++ MapObject.cpp
@@ -130,7 +130,7 @@ g_RenderedObjectsCountInGameWindow++;
 				if (minZ > m_Z)
 					minZ = m_Z;
 
-				g_ZBuffer = (float)minZ;
+				g_ZBuffer = minZ;
 
 				GLfloat tCl = g_DrawColor - 0.1f;
 				glColor3f(tCl, tCl, tCl);
