+++ gump/GumpJournal.cpp
@@ -672,7 +672,7 @@ int TGumpJournal::Draw(bool &mode)
 		g_GL.RestorePort();
 
 		if (g_ShowGumpLocker)
-			g_GL.Draw(g_TextureGumpState[LockMoving], (GLfloat)posX, (GLfloat)posY, 10.0f, 14.0f);
+			g_GL.Draw(g_TextureGumpState[LockMoving], posX, posY, 10.0f, 14.0f);
 	}
 	else
 	{
