+++ gump/GumpPaperdoll.cpp
@@ -606,7 +606,7 @@ int TGumpPaperdoll::Draw(bool &mode)
 		}
 
 		if (g_ShowGumpLocker)
-			g_GL.Draw(g_TextureGumpState[LockMoving], (GLfloat)posX, (GLfloat)posY, 10.0f, 14.0f);
+			g_GL.Draw(g_TextureGumpState[LockMoving], posX, posY, 10.0f, 14.0f);
 	}
 	else
 	{
