+++ gump/GumpSkills.cpp
@@ -658,7 +658,7 @@ int TGumpSkills::Draw(bool &mode)
 		glCallList((GLuint)index);
 
 		if (g_ShowGumpLocker)
-			g_GL.Draw(g_TextureGumpState[LockMoving], (GLfloat)posX, (GLfloat)posY, 10.0f, 14.0f);
+			g_GL.Draw(g_TextureGumpState[LockMoving], posX, posY, 10.0f, 14.0f);
 	}
 	else //ֲûבמנ מבתוךעמג
 	{
