+++ gump/GumpStatusbar.cpp
@@ -627,7 +627,7 @@ int TGumpStatusbar::Draw(bool &mode)
 		glCallList((GLuint)index);
 
 		if (g_ShowGumpLocker)
-			g_GL.Draw(g_TextureGumpState[LockMoving], (GLfloat)posX, (GLfloat)posY, 10.0f, 14.0f);
+			g_GL.Draw(g_TextureGumpState[LockMoving], posX, posY, 10.0f, 14.0f);
 	}
 	else //ֲûבמנ מבתוךעמג
 	{
